feat: implement "up, cd"
